changes to D2VConfig: - move from config.py to new d2v_config.py - add to __init__.py for gamechangerml/configs - update references to this class
