+++ gamechangerml/configs/__init__.py
@@ -6,3 +6,4 @@ from .similarity_config import SimilarityConfig
 from .doc_compare_similarity_config import DocCompareSimilarityConfig
 from .qexp_config import QexpConfig
 from .training_config import TrainingConfig
+from .d2v_config import D2VConfig
